MAINT: rename `pava_pybind` to `pava`
[skip circle]
